Removed ReadonlyApiController TModel context type definition requirement
